+++ CHANGELOG.md
@@ -4,13 +4,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## [v0.4.0pre4] - 2020-05-04
 ### Added
  - Added `-m|--multiple` option to `bot send poll`.
  - Added `--anonymous/--no-anonymous` option to `bot send poll`.
  - Added `--until` option to `bot send poll`.
 
-## [0.4.0pre3] - 2020-04-17
+## [v0.4.0pre3] - 2020-04-17
 ### Changed
  - `--horizontal/-h` and `--vertical/-v` flags for `bot send video` has been
  changed to `--width/-w` ad `--height/-h`.